fix to_absolute It didn't work well if the was context relative. For example +A+B in the context of +A should be +A+B and not +A+A+B
