+++ Gemfile
@@ -3,7 +3,7 @@ source "http://rubygems.org"
 # Add dependencies to develop your gem here.
 # Include everything needed to run rake, tests, features, etc.
 group :development do
-  gem "rspec", "~> 2.8.0"
+  gem "rspec", "~> 3.5.0"
   gem "rdoc", "~> 3.12"
   gem "bundler", "~> 1.1"
   gem "jeweler", "~> 1.8.3"
