+++ lib/smart_name.rb
@@ -3,8 +3,15 @@
 require 'active_support/configurable'
 require 'active_support/inflector'
 require 'htmlentities'
+require_relative 'smart_name/parts'
+require_relative 'smart_name/variants'
+require_relative 'smart_name/contextual'
 
 class SmartName < Object
+  include Parts
+  include Variants
+  include Contextual
+
   RUBYENCODING = RUBY_VERSION !~ /^1\.8/
   OK4KEY_RE    = RUBYENCODING ? '\p{Word}\*' : '\w\*'
 
@@ -16,8 +23,8 @@ class SmartName < Object
   SmartName.banned_array   = ['/', '~', '|']
   SmartName.var_re         = /\{([^\}]*\})\}/
   SmartName.uninflect      = :singularize
-  SmartName.stabilize      = false       
-  
+  SmartName.stabilize      = false
+
 
   JOINT_RE = Regexp.escape joint
 
@@ -26,7 +33,12 @@ class SmartName < Object
   class << self
     def new obj
       return obj if obj.is_a? self.class
-      str = obj.is_a?(Array) ? obj * joint : obj.to_s
+      str =
+        if obj.is_a?(Array)
+          obj.map { |i| i.to_s } * joint
+        else
+          obj.to_s
+        end
       if (known_name = @@name2nameobject[str])
         known_name
       else
@@ -37,7 +49,7 @@ class SmartName < Object
     def banned_re
       %r{#{ (['['] + banned_array << joint) * '\\' + ']' }}
     end
-    
+
     # Sometimes the core rule "the key's key must be itself" (called "stable" below) is violated
     # eg. it fails with singularize as uninflect method for Matthias -> Matthia -> Matthium
     # Usually that means the name is a proper noun and not a plural.
@@ -57,6 +69,7 @@ class SmartName < Object
   # ~~~~~~~~~~~~~~~~~~~~~~ INSTANCE ~~~~~~~~~~~~~~~~~~~~~~~~~
 
   attr_reader :simple, :parts, :key, :s
+  alias simple? simple
   alias to_s s
 
   def initialize str
@@ -112,185 +125,21 @@ class SmartName < Object
     other_key == key
   end
 
-  # ~~~~~~~~~~~~~~~~~~~ VARIANTS ~~~~~~~~~~~~~~~~~~~
-
-  def simple_key
-    decoded
-      .underscore
-      .gsub(/[^#{OK4KEY_RE}]+/, '_')
-      .split(/_+/)
-      .reject(&:empty?)
-      .map { |key| SmartName.stable_uninflect(key) }
-      .join('_')
+  # @return true if name starts with the same parts as `prefix`
+  def starts_with? prefix
+    start_name = prefix.to_name
+    start_name == self[0, start_name.length]
   end
+  alias_method :start_with?, :starts_with?
 
-  def url_key
-    @url_key ||= part_names.map do |part_name|
-      stripped = part_name.decoded.gsub(/[^#{OK4KEY_RE}]+/, ' ').strip
-      stripped.gsub(/[\s\_]+/, '_')
-    end * self.class.joint
-  end
-
-  def safe_key
-    @safe_key ||= key.tr('*', 'X').tr self.class.joint, '-'
-  end
-
-  def decoded
-    @decoded ||= s.index('&') ? HTMLEntities.new.decode(s) : s
-  end
-
-  # ~~~~~~~~~~~~~~~~~~~ PARTS ~~~~~~~~~~~~~~~~~~~
-
-  alias simple? simple
-  def junction?
-    !simple?
-  end
-
-  def left
-    @left ||= simple? ? nil : parts[0..-2] * self.class.joint
-  end
-
-  def right
-    @right ||= simple? ? nil : parts[-1]
-  end
-
-  def left_name
-    @left_name ||= left && self.class.new(left)
-  end
-
-  def right_name
-    @right_name ||= right && self.class.new(right)
-  end
-
-  # Note that all names have a trunk and tag,
-  # but only junctions have left and right
-
-  def trunk
-    @trunk ||= simple? ? s : left
-  end
-
-  def tag
-    @tag ||= simple? ? s : right
-  end
-
-  def trunk_name
-    @trunk_name ||= simple? ? self : left_name
-  end
-
-  def tag_name
-    @tag_name ||= simple? ? self : right_name
-  end
-
-  def part_names
-    @part_names ||= parts.map(&:to_name)
-  end
-
-  def piece_names
-    @piece_names ||= pieces.map(&:to_name)
-  end
-
-  def pieces
-    @pieces ||=
-      if simple?
-        [self]
-      else
-        junction_pieces = []
-        parts[1..-1].inject parts[0] do |left, right|
-          piece = [left, right] * self.class.joint
-          junction_pieces << piece
-          piece
-        end
-        parts + junction_pieces
-      end
-  end
-
-  # ~~~~~~~~~~~~~~~~~~~~ SHOW / ABSOLUTE ~~~~~~~~~~~~~~~~~~~~
-
-  def to_show *ignore
-    ignore.map!(&:to_name)
-
-    show_parts = parts.map do |part|
-      reject = (part.empty? || (part =~ /^_/) || ignore.member?(part.to_name))
-      reject ? nil : part
-    end
-
-    show_name = show_parts.compact.to_name.s
-
-    case
-    when show_parts.compact.empty? then  self
-    when show_parts[0].nil?        then  self.class.joint + show_name
-    else show_name
-    end
-  end
-
-  def to_absolute context, args={}
-    context = context.to_name
-    parts.map do |part|
-      new_part =
-        case part
-        when /^_user$/i
-          name_proc = self.class.session
-          name_proc ? name_proc.call : part
-        when /^_main$/i            then self.class.params[:main_name]
-        when /^(_self|_whole|_)$/i then context.s
-        when /^_left$/i            then context.trunk
-          # note - inconsistent use of left v. trunk
-        when /^_right$/i           then context.tag
-        when /^_(\d+)$/i
-          pos = $~[1].to_i
-          pos = context.length if pos > context.length
-          context.parts[pos - 1]
-        when /^_(L*)(R?)$/i
-          l_s, r_s = $~[1].size, !$~[2].empty?
-          l_part = context.nth_left l_s
-          r_s ? l_part.tag : l_part.s
-        # when /^_/
-        #   custom = args[:params] ? args[:params][part] : nil
-        #   custom ? CGI.escapeHTML(custom) : part
-        # why are we escaping HTML here?
-        else
-          part
-        end.to_s.strip
-      new_part.empty? ? context.to_s : new_part
-    end * self.class.joint
-  end
-
-  def to_absolute_name *args
-    self.class.new to_absolute(*args)
-  end
-
-  def nth_left n
-    # 1 = left; 2= left of left; 3 = left of left of left....
-    (n >= length ? parts[0] : parts[0..-n - 1]).to_name
+  # @return true if name has a chain of parts that equals `subname`
+  def include? subname
+    subkey = subname.to_name.key
+    joint = Regexp.quote self.class.joint
+    key =~ /(^|#{joint})#{Regexp.quote subkey}($|#{joint})/
   end
 
   # ~~~~~~~~~~~~~~~~~~~~ MISC ~~~~~~~~~~~~~~~~~~~~
-
-  def replace_part oldpart, newpart
-    oldpart = oldpart.to_name
-    newpart = newpart.to_name
-    if oldpart.simple?
-      if simple?
-        self == oldpart ? newpart : self
-      else
-        parts.map do |p|
-          oldpart == p ? newpart.to_s : p
-        end.to_name
-      end
-    elsif simple?
-      self
-    else
-      if oldpart == parts[0, oldpart.length]
-        if length == oldpart.length
-          newpart
-        else
-          (newpart.parts + parts[oldpart.length..-1]).to_name
-        end
-      else
-        self
-      end
-    end
-  end
 
   # HACK. This doesn't belong here.
   # shouldn't it use inclusions???
